moved adding deprecation notices into dependency_snapshot
